Fix two _readString bugs causing read corruption.
Fixed two bugs in _readString that caused it to fail to find the end of
a string if the end quote wasn't in the first buffer.

First, the call to refresh the buffer did not take into account that we
were scanning forward, not removing anything, so ensure(1) didn't
necessarily have to do anything.  (In practice it added 1 byte each
time.)

Second, there was a micro-optimization (red flag!) of copying
this.buffer to a local variable *outside* of a loop that called ensure.
The ensure method *replaces* this.buffer since it is an immutable
string.